+++ app.js
@@ -1,17 +1,6 @@
-const validator  = require("validator");
 const chalk = require("chalk");
-const getNotes = require("./utils/notesUtils");
+const getNotes = require("./utils/notesUtils.js");
 
-const blueText = function(text) {
-    return chalk.blue(text);
-}
+const commandArg = process.argv[2];
 
-const redText = function(text) {
-    return chalk.red(text);
-}
-
-
-console.log(getNotes());
-console.log(validator.isEmail("abc@abc")? blueText("Valid Email !!"): redText("WRONGO!"));
-
-console.log("PROCESS ARGV", process.argv);
+console.log(chalk.blue.bold(commandArg));
